20000006-1826 O peso deve permitir valor "0"

+++ Src/MSTech.GestaoEscolar.Entities/Abstracts/Abstract_CLS_QuestionarioResposta.cs
@@ -72,7 +72,7 @@ namespace MSTech.GestaoEscolar.Entities.Abstracts
 		/// Peso da resposta para respostas que compoem testes..
 		/// </summary>
 		[MSNotNullOrEmpty("[qtr_peso] é obrigatório.")]
-		public virtual int qtr_peso { get; set; }
+		public virtual decimal qtr_peso { get; set; }
 
     }
 }
